Adding Initial Auto-Tracking Functionality
This commit adds to the Tracker for Diddy Kong Racing the ability to
perform Auto-Tracking for a good portion of the DKR AP Randomizer. This
addition tracks the Balloons (total and for each separate
World/Overworld). The Keys, and even the settings for the AP Seed it is
tracking. Current functionality not a part of the Auto-Tracking but will
be completed when time presents itself is for Boss Completion, Trophy
Completion, and Tracking when the First Phase of a Boss is Completed for
Logic.